+++ events/events.cpp
@@ -6,52 +6,48 @@
 
 #include "input.h"
 #include "gui.h"
-#include "r_graphics.h"
+#include "r_window.h"
 
 #include "log.h"
 
 using namespace Fresa;
 
 //---Event handling---
-//      This part is very WIP and will soon be replaced by a propper event handler
+//      Here all the system events are processed with SDL built in event system, and they are translated to fresa's event system
 
-Events::EventTypes Events::handleEvents() {
+void Events::handleSystemEvents() {
     SDL_Event event;
     while (SDL_PollEvent(&event)) {
         switch (event.type) {
             case SDL_QUIT:
-                return EVENT_QUIT;
+                event_quit.publish();
             case SDL_KEYDOWN:
-                if (event.key.repeat == 0) {
-                    Input::onKeyDown((Input::Key)event.key.keysym.scancode);
-#ifndef DISABLE_GUI
-                    Gui::addInputKey(event.key.keysym.sym);
-#endif
-                }
+                if (event.key.repeat == 0)
+                    Input::event_key_down.publish((Input::Key)event.key.keysym.sym);
                 break;
             case SDL_KEYUP:
                 if (event.key.repeat == 0)
-                    Input::onKeyUp((Input::Key)event.key.keysym.scancode);
+                    Input::event_key_up.publish((Input::Key)event.key.keysym.sym);
                 break;
             case SDL_MOUSEMOTION:
-                Input::onMouseMove(event.motion.x, event.motion.y); //This event consumes 5-10% of CPU, check in the future
+                Input::event_mouse_move.publish(Vec2<float>(event.motion.x, event.motion.y)); //This event consumes 5-10% of CPU, check in the future
                 break;
             case SDL_MOUSEBUTTONDOWN:
-                Input::onMouseDown(event.button.button);
+                Input::event_mouse_down.publish((Input::MouseButton)event.button.button);
                 break;
             case SDL_MOUSEBUTTONUP:
-                Input::onMouseUp(event.button.button);
+                Input::event_mouse_up.publish((Input::MouseButton)event.button.button);
                 break;
             case SDL_MOUSEWHEEL:
-                Input::onMouseWheel(event.wheel.y);
+                Input::event_mouse_wheel.publish(event.wheel.y);
                 break;
             case SDL_WINDOWEVENT:
                 if (event.window.event == SDL_WINDOWEVENT_FOCUS_LOST)
-                    return EVENT_PAUSE;
+                    event_paused.publish(true);
                 if (event.window.event == SDL_WINDOWEVENT_FOCUS_GAINED)
-                    return EVENT_CONTINUE;
+                    event_paused.publish(false);
                 if (event.window.event == SDL_WINDOWEVENT_SIZE_CHANGED)
-                    Graphics::onResize(Vec2<>(event.window.data1, event.window.data2));
+                    Graphics::Window::event_window_resize.publish(Vec2<>(event.window.data1, event.window.data2));
                 break;
             case SDL_USEREVENT:
                 bool* done = reinterpret_cast<bool*>(event.user.data1);
@@ -59,5 +55,4 @@ Events::EventTypes Events::handleEvents() {
                 break;
         }
     }
-    return EVENT_NONE;
 }