+++ graphics/r_window.h
@@ -6,9 +6,12 @@
 
 #include "config.h"
 #include "r_dtypes.h"
+#include "events.h"
 
 namespace Fresa::Graphics::Window
 {
     WindowData create(Vec2<> size, str name);
     ui16 getRefreshRate(WindowData &win, bool force = false);
+    
+    inline Events::Event<Vec2<>> event_window_resize;
 }